requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
psycopg2-binary==2.9.10
qdrant-client==1.13.2
mem0ai[graph]==0.1.48
python-dotenv==1.0.1
google-generativeai==0.8.4